zephyr-bindgen: split build.rs into its own project
Make the implementation of zephyr-sys's build.rs be a separate cargo
project to avoid having any build-dependencies in zephyr-sys.  This is
to work around buggy behavior with cargo when build libstd
