+++ rust-app/zephyr-sys/zephyr-bindgen/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "zephyr-bindgen"
+version = "0.1.0"
+authors = ["Steven Walter <stevenrwalter@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bindgen = "0.49.0"
